+++ link-section/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `MovableRef` and `Ref` implement `Debug` and `Display` if the referenced type does.
 
+### Changed
+
+- Deprecated `reference` module in favor of `Ref` from the crate root.
+
 ## [0.17.0] - 2026-05-16
 
 ### Added